requirements.txt
flask>=2.0
ultralytics>=8.0
opencv-python
Pillow
numpy
requests
